[neptune-pde-solver] major reconstruct the whole project.
**IMPORTANT UPDATE**
This commit change a lot code in project. Remove some layer, because I
found out that one layer of IR is enough for describe the problem.

Current design is letting the solver focusing on the solving optimization,
pushing the hard work back to the frontend, which is driven by Sympy.

Now there is only one Dialect which is called NeptuneIR. It would implement
some basic arith, which is enough for solving a discrete form of PDE.